Split unique_job middleware test by service
